% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bezirksgebiet}
\alias{bezirksgebiet}
\title{District border (Polygons)}
\format{An \code{sf} object containing polygon data with four features.
\describe{
  \item{NAME}{Name of the country}
  \item{BEZIRKSNUM}{Official district number}
  \item{KANTONSNUM}{Official number of canton}
  \item{EINWOHNERZ}{Number of inhabitants}
  \item{BEZIRKSFLA}{Area}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
bezirksgebiet
}
\description{
Manipulated data from the Swissboundries3D dataset by \href{http://swisstopo.ch}{swisstopo.}
Manipulation: 
\enumerate{
  \item Removed z-Values
  \item Dropped some columns
}
}
\keyword{datasets}
