% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gemeindegebiet}
\alias{gemeindegebiet}
\title{Municipality border (Polygons)}
\format{An \code{sf} object containing polygon data with four features.
\describe{
  \item{NAME}{Name of the country}
  \item{BEZIRKSNUM}{Official district number}
  \item{KANTONSNUM}{Official number of canton}
  \item{BFS_NUMMER}{Number of the Swiss official commune register}
  \item{EINWOHNERZ}{Number of inhabitants}
  \item{GEM_FLAECH}{Area}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
gemeindegebiet
}
\description{
Manipulated data from the Swissboundries3D dataset by \href{http://swisstopo.ch}{swisstopo.}
Manipulation: 
\enumerate{
  \item Removed z-Values
  \item Dropped some columns
  \item Singlepart to multipart (\code{\link[sf]{summarise.sf}} corresponds to
        \code{\link[sf]{st_union}}?)
}
}
\keyword{datasets}
