% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kantonsgebiet}
\alias{kantonsgebiet}
\title{Canton border (Polygons)}
\format{An \code{sf} object containing polygon data with four features.
\describe{
  \item{NAME}{Name of the country}
  \item{KANTONSNUM}{Official number of canton}
  \item{EINWOHNERZ}{Number of inhabitatns (probabbly sum of all multipart features)}
  \item{KANTONSFLA}{Area (probabbly sum of all multipart features)}
  \item{Abk.}{Official abbreviation}
  \item{Beitritt}{Member in the swiss confederation since}
  \item{Hauptort}{Capital}
  \item{Amtssprache}{Official language}
  \item{Franzoesisch}{Name in French}
  \item{Italienisch}{Name in Italian}
  \item{Raetoromanisch}{Name in Romansh}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
kantonsgebiet
}
\description{
Manipulated data from the Swissboundries3D dataset by \href{http://swisstopo.ch}{swisstopo.}
Manipulation: 
\enumerate{
  \item Removed z-Values
  \item Dropped some columns
  \item Singlepart to multipart (summarise.sf = st_union?)
  \item Added some metadata (Abk. and following)
}
}
\keyword{datasets}
