% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{landesgebiet}
\alias{landesgebiet}
\title{Country border (Polygons)}
\format{An \code{sf} object containing polygon data with four features.
\describe{
  \item{NAME}{Name of the country}
  \item{EINWOHNERZ}{Number of Inhabitants}
  \item{LANDESFLAE}{Area}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
landesgebiet
}
\description{
Manipulated data from the Swissboundries3D dataset by \href{http://swisstopo.ch}
{swisstopo.}
}
\details{
Manipulation: 
\enumerate{
  \item Removed z-Values
  \item Dropped some columns
}
}
\keyword{datasets}
