% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{seen}
\alias{seen}
\title{Lakes (1:500'000 Swiss Vector Map)}
\format{An \code{sf} object containing polygon data with following attributes
\describe{
  \item{NAMN1}{Name of the Lake in the respective official language (see NLN1)}
  \item{GROESSTE_S}{Probabbly greatest depth}
  \item{SEESPIEGEL}{Probabbly elevelation of lake level}
  \item{NLN1}{Language}
  \item{SHP_AREA}{Area in m2}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/en/products/maps/national/digital/smv500}
}
\usage{
seen
}
\description{
Manipulated data from the Swiss Map Vector 500 dataset by 
\href{http://swisstopo.ch}{swisstopo}.
}
\details{
Manipulation: 
\enumerate{
  \item Dropped some columns
  \item selected lakes \emph{within} switzland (\code{\link[sf]{st_intersects}})
}
}
\keyword{datasets}
